// src/utils/enums.tsx
/**
 * Enum representing different positions.
 */
export enum Position {
  /**
   * Represents the left position.
   */
  LEFT = 'left',

  /**
   * Represents the right position.
   */
  RIGHT = 'right',
}

/**
 * Enum representing different floating icon press actions.
 */
export enum FloatingIconPressActions {
  /**
   * Represents the opening action.
   */
  OPENING = 'opening',

  /**
   * Represents the closing action.
   */
  CLOSING = 'closing',
}